helper data file, moved data that was in basketballdata, remove @types/axios
