+++ package.json
@@ -10,7 +10,6 @@
     "test": "vitest"
   },
   "dependencies": {
-    "@types/axios": "^0.14.0",
     "axios": "^1.3.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
