+++ src/types/basketballdata.ts
@@ -120,20 +120,3 @@ export type {
     playerInfoDataType,
     teamObject,
 };
-
-const emptyDateObject = {
-    dateObj: true,
-    date: null,
-    home_team: {},
-    home_team_score: 0,
-    id: 0,
-    period: 0,
-    postseason: 0,
-    season: 0,
-    status: 0,
-    time: 0,
-    visitor_team: 0,
-    visitor_team_score: 0,
-};
-
-export default emptyDateObject;
